Move `mergeGlobalConfig`
